### 2017-12-08 #168 @sebastian-fox
fingertipsR
The package provides an interface to the API for the [Fingertips](http://fingertips.phe.org.uk/) website, a site developed by Public Health England. It enables users to import data on over 2,000 indicators of public health in England at different levels of geography.

```
Package: fingertipsR
Version: 0.1.3.9000
Title: Fingertips Data for Public Health
Description: Fingertips (<http://fingertips.phe.org.uk/>) contains data for many indicators of public health in England. The underlying data is now more easily accessible by making use of the API.
Authors@R: c(person("Sebastian", "Fox", , "sebastian.fox@phe.gov.uk", c("aut", "cre")),
    person("Julian", "Flowers", , "julian.flowers@phe.gov.uk", c("aut","ctb")))
    httr,
    DT,
    miniUI,
    shiny,
    shinycssloaders,
    readr
    pander,
    curl
```

https://github.com/PublicHealthEngland/fingertipsR

Data retrieval because it provides users with functions to import data for over 2,000 indicators of public health in England from the [Fingertips](http://fingertips.phe.org.uk/) website.

The target audience is anyone with an interest in public health. This includes researchers, public and private sector workers and the general public. The website gets almost one million hits in a year. The majority of users are in the public sector who need to make decisions that affect the public's health. This website helps inform those decisions. Increasingly, the website is being accessed by academic institutions. The *fingertipsR* package is already on CRAN and it is being discussed in some institutions as a future resource for teaching materials.

No

 NA

- [Already on CRAN] Do you intend for this package to go on CRAN?
- [No] Do you wish to automatically submit to the [Journal of Open Source Software](http://joss.theoj.org/)? If so:
- [No] Do you wish to submit an Applications Article about your package to [Methods in Ecology and Evolution](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? If so:
	- [ ] You intend to archive the code for the package in a long-term repository which meets the requirements of the journal.
