### 2017-12-08 #168 @karthik
👋 sebastian-fox
Thank you for your submission to rOpenSci's onboarding. I've tagged your submission as a `presubmission` while I do some preliminary checks before I assign an editor.
