### 2017-12-13 2682a77
coveralls coverage calculation added to .travis.yml. Some badges added to README file
