+++ NEWS.md
@@ -2,6 +2,10 @@
 
 * modifications to the fingertipsR paper
 
+* badges added to README
+
+* package prepared for submission to ropensci
+
 # fingertipsR 0.1.3
 
 * API structure updated to include 99.8 and 95 confidence intervals. Reflected in the outputs of `fingertips_data`. **NOTE** earlier versions of the package will not work anymore because of the underlying change in the API structure
