+++ .travis.yml
@@ -3,3 +3,9 @@
 language: R
 sudo: false
 cache: packages
+
+r_github_packages:
+  - r-lib/covr
+
+after_success:
+  - Rscript -e 'covr::coveralls()'
