### 2017-12-13 78c0969
more tests created to improve coverage
