+++ tests/testthat/test-extract.R
@@ -6,11 +6,17 @@ context("fingertips data extract")
 df1 <- fingertips_data(IndicatorID = 92309)
 df2 <- suppressWarnings(fingertips_data(IndicatorID = 92309, DomainID = 1938132983))
 df3 <- suppressWarnings(fingertips_data(IndicatorID = 92309, ProfileID = 19))
+df4 <- fingertips_data(DomainID = 1938132767)
+df5 <- fingertips_data(ProfileID = 132)
+df6 <- fingertips_data(IndicatorID = 92309, ProfileID = NA)
+df7 <- fingertips_data(DomainID = 1938132767, AreaCode = "E06000015")
+df8 <- fingertips_data(DomainID = 1938132767, rank = TRUE)
+
+ncols <- 24
 
 test_that("the data returned are the same despite different inputs", {
         expect_equal(df1, df2)
         expect_equal(df1, df3)
-        expect_equal(ncol(df1), 24)
 })
 
 test_that("error messages work", {
@@ -19,4 +25,21 @@ test_that("error messages work", {
                      "Invalid AreaTypeID\\. Use function area_types\\(\\) to see what values can be used\\.")
         expect_error(fingertips_data(IndicatorID = 92309, AreaTypeID = NULL),
                      "AreaTypeID must have a value\\. Use function area_types\\(\\) to see what values can be used\\.")
+        expect_error(fingertips_data(DomainID = 1938132767, AreaCode = "hello"),
+                     "Area code not contained in AreaTypeID\\.")
+})
+
+test_that("warning messages work", {
+        expect_warning(fingertips_data(DomainID = 1938132767, AreaCode = "E06000015", ParentAreaTypeID = 153),
+                       "AreaTypeID not a child of ParentAreaTypeID\\. There may be duplicate values in data\\. Use function area_types\\(\\) to see mappings of area type to parent area type\\.")
+
+})
+
+test_that(paste("number of fields returned by fingertips_data function are", ncols), {
+        expect_equal(ncol(df1), ncols)
+        expect_equal(ncol(df4), ncols)
+        expect_equal(ncol(df5), ncols)
+        expect_equal(ncol(df6), ncols)
+        expect_equal(ncol(df7), ncols)
+        expect_equal(ncol(df8), ncols + 3)
 })
