### 2017-12-13 bb531ad
fixed indicator_metadata parsing method
