### 2017-12-15 95f6409
README.md with coverage badge included
