### 2017-12-15 4e882ae
minor changes for deprivation decile documentation and function for GP area type
