### 2017-12-18 #168 @noamross
Reviewers are @carlganz and @nacnudus. In light of the end-of-year holidays, due date is 2018-01-15.
