### 2017-12-18 #168 @noamross
@sebastian-fox You can add our Peer Review badge to your README.  It will read "in-progress"  and update after acceptance:
