### 2017-12-18 64754b7
rOpenSci badge added
