### 2017-12-29 #168 @nacnudus
Hi @sebastian-fox @noamross, here's my review.

By Duncan Garmonsway @nacnudus

- [x] As the reviewer I confirm that there are no conflicts of interest for me to review this work.

- [x] **Performance:** Any performance claims of the software been confirmed (no
    claims made)

---

The fingertipsR package provides programmatic access to data published by Public Health England on the [Fingertips](http://fingertips.phe.org.uk/) website.  This is useful in its own right, but especially because the downloads offered by the website appear to be exclusively as spreadsheets.

The package has some fairly heavy dependencies (both dplyr and data.table), but makes the most of shiny and related packages to provide a searchable index of datasets.

The vignettes demonstrate common usage and are easy to follow, and data structure is self-explanatory once it has been retrieved.

The introductory statement explains that the package is for interacting with a data tool called 'Fingertips'.  It would be helpful if it also explained what the Fingertips data tool is.

The first installation instruction requires the devtools package.  Please could you also give instructions that only use base R, e.g.  https://stackoverflow.com/a/1474125/937932.  Please can you also add an instruction for installing from CRAN, and make it clear that there is a difference, e.g. 'stable' vs 'development'.

Please can the example in the README be described in full.  The current text is generic:

> This is a basic example which shows you how to solve a common problem:

The instruction for browsing vignettes only works if the vignettes have been installed, which is not the default.  If using devtools, then the call should be

```r
devtools::install_github("PublicHealthEngland/fingertipsR",
                         build_vignettes = TRUE,
                         dependencies = "suggests")
```

#### Life Expectancy vignette

I read this vignette before reading any of the function documentation, so some of my points are answered in the docs (but it would still be helpful not to have to go to the docs for the answers).

The `.md` [file](https://github.com/PublicHealthEngland/fingertipsR/blob/master/vignettes/lifeExpectancy.md) committed to GitHub is out of date.

There seems to be a line missing

> We need to begin by calling the fingertipsR package:

Should it be followed by

```r
library(fingertipsR)
```

?

In the code chunk under 'IndicatorID', the comment beginning `# Because the same indicators ...` would be easier to read (and render more nicely as html!) if it were wrapped to lines of 80 characters as below.  It also mentions 'profiles' without defining what profiles are.  That is defined on the Fingertips website, but it would be helpful to repeat some information here for users who have come straight here.  They might come here first if they are maintaining code written by others.  From the [packaging_guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md):

> * If your package connects to a data source or online service, or wraps other software, consider that your package README may be the first point of entry for users.  It should provide enough information for users to understand the nature of the data, service, or software, and provide links to other relevant data and documentation.  For instance, a README should not merely read, "Provides access to GooberDB," but also include, "..., an online repository of Goober sightings in South America.  More information about GooberDB, and documentation of database structure and metadata can be found at *link*.

If filtering the output of `indicators()` is a common workflow (code chunk below), then it might help users if there was a built-in function to return one record per `IndicatorID`.

```r
inds <- indicators()
life_expectancy <- inds[grepl("life expectancy", tolower(inds$IndicatorName)),]

# Because the same indicators are used in multiple profiles, there are many
# repeated indicators in this table (some with varying IndicatorName but same
# IndicatorID)

# This returns a record for each IndicatorID
life_expectancy <-
  unique(life_expectancy[duplicated(life_expectancy$IndicatorID) == FALSE,
                        c("IndicatorID", "IndicatorName")])
knitr::kable(life_expectancy, row.names = FALSE)
```

`areaTypes()$AreaTypeID` is numeric vector, but only contains integers.  Would you consider using an integer vector instead.  Similarly `areaTypes()$ParentAreaTypeID`

Please could you explain the relationship between `AreaTypeID` and `ParentAreaTypeID`.  It is another thing that is probably explained on the Fingertips website, but that would be convenient to know at this point in the vignette.

Typo: `though it change` should be `though it changes`.

The quoted paragraph below begs the question, are there any other defaults, and are they guessable?  What's the rationale?  Also, what do `101`, `102` and `6` translate to?

> ParentAreaTypeID is 6 by default for the fingertips_data() function for AreaTypeID of 102 (though it change if different AreaTypeIDs are entered), so we can stick with that in this example.

Typo in the docs for `deprivation_decile()`.  `Outputs a data frame allocating deprivation decile to area code base on` should be `based on`.

The deprivation section interrupts the workflow from finding the IDs to downloading the data with `fingertips_data()`.  It would flow better to finish downloading the data before seeking the deprivation deciles.

It also took me a while to understand what `deprivation_decile()` does.  It downloads a dataset that is already available in Fingertips the usual way, filters it for the given IDs, renames a column nd recalculaties the decile column.  It would be helpful to explain somewhere why the deprivation decile data gets this special treatment.

Should the error message below be `"Year must be 2010, 2011, 2012 or 2015"`?

```r
    if (!(Year %in% c(2010, 2011, 2012, 2015))) {
        stop("Year must be either 2010 or 2015")
    }
```

The style of the following `if` statement could be improved by simplifying.

```r
    if (!(Year %in% c(2010, 2011, 2012, 2015))) {
        stop("Year must be either 2010 or 2015")
    }
    else if (Year %in% c(2010, 2011, 2012)) {
        IndicatorID <- 338
    }
    else if (Year == 2015) {
        IndicatorID <- 91872
    }
    if (AreaTypeID == 101) {
        if (Year %in% c(2010, 2015)) {
            AreaFilter <- "District & UA"
        }
        else if (Year %in% c(2011, 2012)) {
            stop("Year must be either 2010 or 2015 for AreaTypeID of 101 or 102")
        }
    }
    else if (AreaTypeID == 102) {
        if (Year %in% c(2010, 2015)) {
            AreaFilter <- "County & UA"
        }
        else if (Year %in% c(2011, 2012)) {
            stop("Year must be either 2010 or 2015 for AreaTypeID of 101 or 102")
        }
    }
    else if (AreaTypeID == 7) {
        AreaFilter <- "GP"
    }
    else {
        stop("AreaTypeID must be either 101 (Local authority districts and Unitary Authorities), 102 (Counties and Unitary Authorities) or 7 (General Practice).")
    }
```

Here is a shorter version that I find easier to follow.

```r
    if (!(Year %in% c(2010, 2011, 2012, 2015))) {
        stop("Year must be either 2010 or 2015")
    }
    if (!(AreaTypeID %in% c(101, 102, 7))) {
        stop("AreaTypeID must be either 101 (Local authority districts and Unitary Authorities), 102 (Counties and Unitary Authorities) or 7 (General Practice).")
    }
    if ((AreaTypeID %in% c(101, 102)) && !(Year %in% c(2010, 2015))) {
        stop("Year must be either 2010 or 2015 for AreaTypeID of 101 or 102")
    }
    if (Year %in% c(2010, 2011, 2012)) IndicatorID <- 338
    if (Year == 2015) IndicatorID <- 91872
    if (AreaTypeID == 101) AreaFilter <- "District & UA"
    if (AreaTypeID == 102) AreaFilter <- "County & UA"
    if (AreaTypeID == 7) AreaFilter <- "GP"
```

The output of `pander::pandoc.table()` doesn't work in the HTML rendering of the vignette.  It gives the following output, which is markdown syntax but in a code block instead of being rendered as HTML.

```
##
##
## |  &nbsp;  | IndicatorID |              IndicatorName              | ParentCode |
## |:--------:|:-----------:|:---------------------------------------:|:----------:|
## | **5827** |    90362    | 0.1i - Healthy life expectancy at birth | E12000005  |
## | **5828** |    90362    | 0.1i - Healthy life expectancy at birth | E12000006  |
## | **5829** |    90362    | 0.1i - Healthy life expectancy at birth | E12000008  |
## | **5830** |    90362    | 0.1i - Healthy life expectancy at birth | E12000005  |
## | **5831** |    90362    | 0.1i - Healthy life expectancy at birth | E12000008  |
## | **5832** |    90362    | 0.1i - Healthy life expectancy at birth | E12000005  |
...
```

#### Interactively selecting indicators vignette

Does the word 'significantly' in this sentence imply that a statistical significance test has been applied?  If not then less loaded term may be better, e.g. 'notably', or an explanation of how the indicators are chosen.

> highlighting the areas for each indicator that are deteriorating and are already significantly worse than the England (or parent) value

```r
library(fingertipsR)
library(ggplot2)
inds <- select_indicators()
```

Have you considered wrapping the `geom_tile()` plot as a built-in plot function?  If so, it would be worth refining the design.  Points might be easier to read than tiles.

#### Function documentation

In general it would be helpful to make it clear that the named government departments are in the UK.

```r
area_types
category_types
deprivation_decile
fingertips_data
fingertips_redred
indicator_areatypes
indicator_metadata
indicators
profiles
select_indicators
```

#### `area_types()`

Needs an example that uses the second argument `AreaTypeID`.  This would be a good place to explain how area types relate to parent area types, or at least link to an explanation.

#### `category_types`

Please could it return its result visibly (i.e. `category_types()` in the console should print its output to the console).

Should have an example, even if it seems trivial.

#### `deprivation_decile`

The restriction on the `Year` argument to the years 2010 and 2015 isn't reflected in the code, which also accepts 2011 and 2012.

It would be helpful to add a `Details` section to explain what this function actually does (downloads a dataset the usual way, filters it, renames and recalculates a couple of columns).  Presumably there is a good reason not to just download the deprivation data oneself?

#### `fingertips_data`

The different IDs are a bit overwhelming.  Is there a way to explain how they relate to one another?  Perhaps a drawing would help.

Would you reconsider the `stringsAsFactors` argument?  I am not sure that it is useful for strings to be converted to factors in this case.  If it is useful, then it would be more useful (and less dangerous) if they were converted to a standard sets of factor levels that were guaranteed to included levels that might be missing from the data.  For example, if an area were missing from a particular dataset, it should still have a factor level, so that the data could safely be joined to another dataset of the same areas.

What do the defaults `6`, `101` and `102` translate to?

Typo in the examples: trailing `'#`.

```r
     fingdata <- fingertips_data(DomainID = doms)#'
```

The matter of polarity could be explained more.  I think it's something to do with whether a bigger number is better or worse, but I don't see why that leads to duplicates in the example.

#### `fingertips_redred`

Does the word 'significantly' imply a statistical test is performed?  If not, a less loaded term would be better, e.g. 'notably', or 'quite a bit'.

If `Comparator = "Parent"`, does the spot test still relate to `"England"` as per the text?

The `@return` heading is not being rendered correctly.

#### `indicator_areatypes`

This function loads a local dataset.  Is it not available from the API?  Otherwise it would make more sense (to me!) to explort the dataset than to wrap it in this function.  Though I can see that it is consistent with the other functions that return datasets (and that accept arguments).

#### `indicator_metadata`

Good.

#### `indicators`

Is the title correct?  "Profiles".

#### `profiles`

Good.

#### `select_indicators`

Nice function :)

#### Community guidelines / Code of Conduct

Please could you add guidelines or a code of conduct in the README or a CONTRIBUTING file or a CODE_OF_CONDUCT.md file, and a `Maintainer:` tag in the DESCRIPTION file.  See the
[packaging_guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md).

> We recommend that you use a code of conduct such as the Contributor Covenant in developing your package. You can document your code of conduct in a CODE_OF_CONDUCT.md or CONDUCT.md file in the package root directory, and linking to this file from the README.md file. devtools::use_code_of_conduct() will add the Contributor Covenant template to your package.

#### NEWS

Please follow the [packaging guide](https://github.com/ropensci/onboarding/blob/master/packaging_guide.md) advice on putting CRAN release dates in the NEWS file, and tagging the release on GitHub.

#### DESCRIPTION

If the UK Government owns the copyright on this package then that should be mentioned in the DESCRIPTION.

Please could more tests be added?  Some lines of code are untested so far, as the following code will show.  This is fair enough for the shiny function `select_indicators()`, but would be straightforward to hit 100% of the other functions.  The `select_indicators()` function can be excluded from coverage statistics by using the comments `# nocov start` and `# nocov end`.

```r
x <- covr::package_coverage()
covr::zero_coverage(x)
```

#### goodpractice

There are a lot of long lines, and I am one of those people who likes to fit other windows on the screen besides the editor ;)  Sometimes it's unavoiadable because of the API calls, but in most cases function arguments could flow onto another line, or inline sequences of piped functions could be wrapped in their own function.

It would be well worth asking a copy editor, or someone of that ilk, to edit the docs and vignettes.  It just needs a bit of tidying up to make it easier to read, e.g. changing `Other data extract functions` to `Other data extraction functions`.

`fingertips_redred()`, `indicator_metadata()` return tibbles, which is very nice.  Would you consider also returning tibbles from the other functions?  `area_types()`, `category_types()`, `indicator_areatypes()`, `deprivation_decile()`, `fingertips_data()`, `areatypes_by_indicators`, `indicators()` and `profiles()`?

`area_types` and `areatypes` are both used in function names.  Maybe not worth changing now because backwards compatibility is important and the package is already widely used.

Please can you remove the commented-out code in `retrieve_data.R`.

Not a question about the package, but for Public Health England -- why isn't http://fingertips.phe.org.uk/ on a gov.uk domain?

I hope this review is helpful.  Please ask any questions you have.

#### Time spent reviewing

6 hours
