### 2018-01-03 #168 @carlganz
* Readme should probably include a more detailed statement of need.
      * As @nacnudus stated, `devtools::install_github` should probably include build vignette option.
      * Missing documentation for `indicators2`. What is this function?
      * `select_indicator` should probably have a note about the long loading time
      * Typo https://github.com/PublicHealthEngland/fingertipsR/blob/78c09694d36e7bcda0dd71b05017f556020b5404/R/enhancements.R#L5
      * Missing examples for `category_types`, `indicator_areatypes`; also `category_types` doesn't work
      * Typo  https://github.com/PublicHealthEngland/fingertipsR/blob/master/R/fingertips_data.R#L34
      * Again echoing @nacnudus, please include contributing guidelines, and the copyright holder in the description

    * Echoing @nacnudus, please use code coverage. There seem to be enough tests, but without code coverage it is difficult to tell.

---

The `fingerTipsR` package does an excellent job providing easy access to a substantial amount of data. It even provides a nifty Shiny gadget for interactive selection of data. I think there are a few areas where the package can be improved, but overall the package is high quality.

The readme is a little sparse. Please add a statement of need. A more detailed description of what the Fingertips data source is would be nice. How is the data collected? What is scope of dataset? Ideally, a person who has never heard of the Fingertips dataset should be able to read the readme and then have a basic understanding of what the dataset is.

Vignettes do a good job explaining how to use the package. Echoing @nacnudus flipping the "Deprivation" and "Extracting the data" sections from the "Life expectancy by deprivation" vignette would make the document flow better.

The `indicator2` function has no documentation.

`category_types` and `indicator_areatypes` do not have examples.

For functions that take a long time to run, it would be wise to add a note to documentation saying so.

The Fingertips dataset is pretty substantial. This makes certain functionality pretty slow. There are a few ways the code could be optimized. These are simply suggestions on how to improve the code. I will still recommend approving this package even if you don't incorporate these changes.

* Remove `data.table` dependency

`data.table::rbindlist` is functionally equivalent to `dplyr::bind_rows`. Since `dplyr` is already imported may as well use `bind_rows`

* Remove `purrr::map_df(rbind)`

This is also functionally equivalent to `dplyr::bind_rows`, but is much slower. This would also remove the `purrr` dependency, but I'm about to recommend using `purrr` for something else.

* Don't build data structures in `for` loop

@noamross has a good write up about why this idiom is memory inefficient in [section 4 here](http://www.noamross.net/blog/2014/4/16/vectorization-in-r--why.html)

Here is one example in the package: https://github.com/PublicHealthEngland/fingertipsR/blob/master/R/indicators.R#L42

There are a few different ways of dealing with this. The `purrr` based solution would look something like this:

df <- DomainId %>% map_dfr(function(dom) {
          dfRaw <- paste0(path,"indicator_metadata/by_group_id?group_ids=",dom) %>%
                        GET %>%
                        content("text") %>%
                        fromJSON(flatten = TRUE)
                if (length(dfRaw) != 0){
                        dfRaw <- unlist(dfRaw, recursive = FALSE)
                        dfIDs <- dfRaw[grepl("IID", names(dfRaw))]
                        names(dfIDs) <- gsub(".IID","",names(dfIDs))
                        dfDescription <- unlist(dfRaw[grepl("Descriptive", names(dfRaw))],
                                                recursive = FALSE)
                        dfDescription <- dfDescription[grepl("NameLong", names(dfDescription))]
                        names(dfDescription) <- gsub(".Descriptive.NameLong","",names(dfDescription))
                        commonNames <- intersect(names(dfIDs), names(dfDescription))
                        dfIDs <- dfIDs[commonNames]
                        dfDescription <- dfDescription[commonNames]

                        data.frame(IndicatorID = unlist(dfIDs),
                                         IndicatorName = unlist(dfDescription),
                                         DomainID = dom,
                                         row.names=NULL)
                }
})

```

This could also be done with `lapply`.

* Fix `category_types`

Doesn't return anything.

Hopefully this review was helpful.

Carl Ganz
