### 2018-01-03 #168 @noamross
Thank you for your thorough reviews, @nacnudus and @carlganz!  @sebastian-fox, at this point we generally aim for a response to reviewers within two weeks. This can be fully implementing recommended changes or just a written response of the changes you plan to make and a time frame ([example](https://github.com/ropensci/onboarding/issues/87#issuecomment-312666878)).

I would re-emphasize @nacnudus's comment on thinking about the README and other documentation as first points of entry for users.  We try to follow this mentality throughout package documentation.  I also strongly suggest you follow @carlganz's recommendations on performance and reducing dependencies.
