### 2018-01-03 bc0453c
Final (?) changes from simon.
Minor changes to text
