+++ vignettes/fingertipsR_paper.Rmd
@@ -31,7 +31,7 @@ The indicators are arranged in thematic or topical profiles covering a wide rang
 The data can be accessed from <https://fingertips.phe.org.uk> where the data can be visualised using a variety of methods.
 Data can be also be exported for a group of indicators, for a single geography per download. 
 
-A recent addition to the Fingertips platform was an [Automated Programming Interface](https://fingertips.phe.org.uk/api) (API) to enable developers to re-use the data. 
+A recent addition to the Fingertips platform was an [Application Programming Interface](https://fingertips.phe.org.uk/api) (API) to enable developers to re-use the data. 
 To facilitate access to Fingertips data we have developed an R package - `fingertipsR` - enabling rapid and easy access to fingertips data by analysts and data scientists. 
 The package is available from the [Comprehensive R Archive Network](https://cran.r-project.org/) (CRAN).
 
@@ -48,9 +48,10 @@ This paper describes the Fingertips API and the use of the `fingertipsR` package
 Data are organised thematically, grouping related indicators together. 
 The web interface provides an interactive platform for data visualisations allowing users to examine trends over time, geographical distribution and make comparisons between health providers. 
 The web interface also provides a means to download the data behind the visualisations for re-use.
-However, accessing the data in this manner limits the user to a single group of indicators for a specified geography per download. 
+However, accessing the data in this manner limits the user to a single group of indicators within for a specified geography per download, creating a time burden on the researchers. 
 To enable programmatic access to the data PHE developed an application programming interface (API). 
-Use of an API is highly technical and not always suited to the public health researcher. 
+In contrast to the web interface, the API allows a researcher to extract data for many different indicators across domains and for multiple geographies. 
+However, use of an API is highly technical and not always suited to the public health researcher. 
 
 R is a free, open source software for statistical analysis. [@rsoftware]
 It doubles as both a programming language and analytical environment for performing statistical analyses.
@@ -71,12 +72,12 @@ The `fingertipsR` package extends R by providing an easy-to-use set of functions
 
 Public health data gathered and synthesised by PHE are stored on Fingertips in nested thematic groups. 
 Profiles group together broad themes of data such as antimicrobial resistance or diabetes. [@johnson2016improving]
-These profiles may consist of multiple domains - such as prevalence or targets.
-Individual indicators then provide actual values for different measures within the domains for example prevalence of dementia among those $\geq$ 65 years of age. 
+These profiles may consist of multiple domains bundling together common indicators within a theme - such as prevalence of disease or primary care treatment targets. <!-- e.g. diabetes profile -->
+Individual indicators then provide actual values for different measures within the domains, for example prevalence of dementia among those $\geq$ 65 years of age. 
 In addition, indicators can vary by different area types within two broad geographic themes, administrative and health.
 Health geographies include commissioners of health care services for local areas (known as Clinical Commissioning Groups or CCGs), hospital groups (acute trusts) and general practices. 
 Administrative geographies include upper- and lower-tier local authorities and government regions among others.
-These geographies themselves fall within nested geographies based on whether they are in the administrative or health geography hierarchy. 
+These geographies themselves fall within nested geographies according to whether they are in the administrative or health geography hierarchy. 
 
 All indicators have a fundamentally identical structure. 
[…]
