### 2018-01-11 9f3b1ad
README updated with improved example, added rOpenSci badge, improved statement of Fingertips tool, and build_vignettes added to install_github
