+++ README.md
@@ -7,7 +7,7 @@
 fingertipsR
 ===========
 
-This is an R package to interact with Public Health England's [Fingertips](http://fingertips.phe.org.uk/) data tool. This can be used to load data from the Fingertips API into R for further manipulation.
+This is an R package to interact with Public Health England's [Fingertips](http://fingertips.phe.org.uk/) data tool. Fingertips is a major public repository of population and public health indicators for England. The site presents the information in many ways to improve accesibility for a wide range of audiences ranging from public health professionals and researchers to the general public. This package can be used to load data from the Fingertips API into R for further use.
 
 Installation
 ------------
@@ -27,32 +27,103 @@ You can install the latest version of fingertipsR from github with:
 
 ``` r
 # install.packages("devtools")
-devtools::install_github("PublicHealthEngland/fingertipsR")
+devtools::install_github("PublicHealthEngland/fingertipsR",
+                         build_vignettes = TRUE,
+                         dependencies = "suggests")
 ```
 
 Example
 -------
 
-This is a basic example which shows you how to solve a common problem:
+This is an example of a workflow for downloading data for the indicator on *Healthy Life Expectancy at Birth* from the *Public Health Outcomes Framework* profile.
+
+The `profiles()` function presents all of the available profiles:
 
 ``` r
 library(fingertipsR)
-df <- profiles(ProfileName = "Public Health Outcomes Framework")
-print(df)
-#>   ProfileID                      ProfileName   DomainID
-#> 1        19 Public Health Outcomes Framework    1000049
-#> 2        19 Public Health Outcomes Framework    1000041
-#> 3        19 Public Health Outcomes Framework    1000042
-#> 4        19 Public Health Outcomes Framework    1000043
-#> 5        19 Public Health Outcomes Framework    1000044
-#> 6        19 Public Health Outcomes Framework 1938132983
-#>                           DomainName
-#> 1             Overarching indicators
-#> 2       Wider determinants of health
-#> 3                 Health improvement
-#> 4                  Health protection
-#> 5 Healthcare and premature mortality
-#> 6             Supporting information
+profs <- profiles()
+profs <- profs[grepl("Public Health Outcomes Framework", profs$ProfileName),]
+head(profs)
+#>    ProfileID                      ProfileName   DomainID
+#> 22        19 Public Health Outcomes Framework    1000049
+#> 23        19 Public Health Outcomes Framework    1000041
+#> 24        19 Public Health Outcomes Framework    1000042
+#> 25        19 Public Health Outcomes Framework    1000043
+#> 26        19 Public Health Outcomes Framework    1000044
+#> 27        19 Public Health Outcomes Framework 1938132983
+#>                            DomainName
+#> 22             Overarching indicators
+#> 23       Wider determinants of health
+#> 24                 Health improvement
+#> 25                  Health protection
+#> 26 Healthcare and premature mortality
+#> 27             Supporting information
+```
+
+This table shows that the `ProfileID` for the Public Health Outcomes Framework is 19. This can be used as an input for the `indicators()` function:
+
+``` r
+profid <- 19
+inds <- indicators(ProfileID = profid)
+print(inds[grepl("Healthy", inds$IndicatorName), c("IndicatorID", "IndicatorName")])
+#>     IndicatorID
+#> 122       92543
+#> 168       90362
+#>                                                                                                                                                                                            IndicatorName
+#> 122                                                                               2.05ii - Proportion of children aged 2-2½yrs offered ASQ-3 as part of the Healthy Child Programme or integrated review
[…]
