### 2018-01-11 4b958ac
README improved install package instructions
