### 2018-01-11 a98b62b
added indicators_unique and small changes to LE vignette based on rOpenSci feedback
