+++ R/indicator_metadata.R
@@ -23,7 +23,8 @@
 #'   \code{\link{deprivation_decile}} for deprivation lookups and
 #'   \code{\link{area_types}} for area types and their parent mappings and
 #'   \code{\link{category_types}} for category lookups and
-#'   \code{\link{indicator_areatypes}} for indicators by area types lookups
+#'   \code{\link{indicator_areatypes}} for indicators by area types lookups and
+#'   \code{\link{indicators_unique}} for unique indicatorids and their names
 #' @export
 
 indicator_metadata <- function(IndicatorID = NULL,
