+++ R/profiles.R
@@ -21,7 +21,10 @@
 #' @seealso \code{\link{area_types}} for area type  and their parent mappings,
 #'   \code{\link{indicators}} for indicator lookups,
 #'   \code{\link{indicator_metadata}} for indicator metadata and
-#'   \code{\link{deprivation_decile}} for deprivation decile lookups
+#'   \code{\link{deprivation_decile}} for deprivation decile lookups and
+#'   \code{\link{category_types}} for category lookups and
+#'   \code{\link{indicator_areatypes}} for indicators by area types lookups and
+#'   \code{\link{indicators_unique}} for unique indicatorids and their names
 #' @export
 
 profiles <- function(ProfileID = NULL, ProfileName = NULL) {
