+++ vignettes/lifeExpectancy.Rmd
@@ -1,7 +1,5 @@
 ---
 title: "Plotting healthy life expectancy and life expectancy by deprivation for English local authorities"
-author: "Seb Fox"
-date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{Life expectancy by deprivation}
@@ -11,7 +9,13 @@ vignette: >
 
 This worked example attempts to document a common workflow a user might follow when using the `fingertipsR` package.
 
-Suppose you want to plot healthy life expectancy and life expectancy by deprivation for a given year of data that fingertips contains - you will begin by wondering *where to start*.
+`fingertipsR` provides users the ability to import data from the [Fingertips](http://fingertips.phe.org.uk/) website. Fingertips is a major repository of public health indicators in England. The site is structured in the following way:
+
+* Profiles - these contain indicators related to a broad theme (such as a risk factor or a disease topic etc)
+* Domains - these are subcategories of profiles, and break the profiles down to themes within the broader theme
+* Indicators - this is the lowest level of the structure and sit within the domains. Indicators are presented at different time periods, geographies, sexes, ageband and categories.
+
+This example demonstrates how you can plot healthy life expectancy and life expectancy by deprivation for a given year of data that fingertips contains. So, *where to start*?
 
 ## Where to start
 
@@ -33,9 +37,8 @@ Therefore, the inputs to the `fingertips_data` function that we need to find out
 * ParentAreaTypeID
 
 We need to begin by calling the `fingertipsR` package: 
-```{r packages, echo=FALSE}
+```{r packages}
 library(fingertipsR)
-library(ggplot2)
 ```
 
 ## IndicatorID
@@ -51,7 +54,9 @@ We can use the `indicators()` function to return a list of all the indicators wi
 inds <- indicators()
 life_expectancy <- inds[grepl("life expectancy", tolower(inds$IndicatorName)),]
 
-# Because the same indicators are used in multiple profiles, there are many repeated indicators in this table (some with varying IndicatorName but same IndicatorID)
+# Because the same indicators are used in multiple profiles, there are many
+# repeated indicators in this table (some with varying IndicatorName but same
+# IndicatorID)
 
 # This returns a record for each IndicatorID
 life_expectancy <- unique(life_expectancy[duplicated(life_expectancy$IndicatorID) == FALSE,
@@ -129,9 +134,10 @@ DT::datatable(data, filter = "top", rownames = FALSE) #note, this line will only
 
 ## Plotting outputs
 
-Using `ggplot2` it is possible to plot the outputs
+Using `ggplot2` it is possible to plot the outputs.
 
-```{r plot, fig.width=8, fig.height=5}
+```{r plot, fig.width=8, fig.height=5, message=FALSE}
+library(ggplot2)
 p <- ggplot(data, aes(x = IMDscore, y = Value, col = factor(IndicatorID)))
 p <- p + 
         geom_point() +
