### 2018-01-11 ad76547
converted output of numeric area_types fields to integer
