### 2018-01-11 a293be7
deprivation_decile function check statements changed to be more readable. Changes to LE vignette (more explanation for people unfamiliar with ftips, and moved deprivation decile section around). Minor changes to SI vignette
