+++ vignettes/lifeExpectancy.Rmd
@@ -28,7 +28,7 @@ There is one function in the `fingertipsR` package that extracts data from the F
 * AreaTypeID (this defaults to 102; County and Unitary Authority)
 * ParentAreaTypeID (this defaults to 6 for an AreaTypeID of 102; Government Office Region)
 
-One of *IndicatorID*, *DomainID* or *ProfileID* must be complete. *AreaCode* needs completion if you are extracting data for a particular area or group of areas only. *AreaTypeID* determines the geography to extract the data for. In this case we want County and Unitary Authority level. *ParentAreaTypeID* requires an area type code that the *AreaTypeID* maps to, though if left out one will be chosen automatically.
+At least one of *IndicatorID*, *DomainID* or *ProfileID* must be complete. These fields relate to each other as described in the introduction. *AreaCode* needs completing if you are extracting data for a particular area or group of areas only. *AreaTypeID* determines the geography to extract the data for. In this case we want County and Unitary Authority level. *ParentAreaTypeID* requires an area type code that the *AreaTypeID* maps to at a higher level of geography. For example, when combining groups of County and Unitary Authorities it is possible to create Goverenment Office Regions. These mappings can be identified using the `area_types()` funciton. If ignored, a *ParentAreaTypeID* will be chosen automatically.
 
 Therefore, the inputs to the `fingertips_data` function that we need to find out are the ID codes for:
 
@@ -51,16 +51,9 @@ There are two indicators we are interested in for this exercise. Without consult
 We can use the `indicators()` function to return a list of all the indicators within Fingertips. We can then filter the name field for the term *life expectancy* (note, the IndicatorName field has been converted to lower case in the following code chunk to ensure matches will not be overlooked as a result of upper case letters).
 
 ```{r indicators}
-inds <- indicators()
+inds <- indicators_unique()
 life_expectancy <- inds[grepl("life expectancy", tolower(inds$IndicatorName)),]
 
-# Because the same indicators are used in multiple profiles, there are many
-# repeated indicators in this table (some with varying IndicatorName but same
-# IndicatorID)
-
-# This returns a record for each IndicatorID
-life_expectancy <- unique(life_expectancy[duplicated(life_expectancy$IndicatorID) == FALSE,
-                                          c("IndicatorID", "IndicatorName")]) 
 knitr::kable(life_expectancy, row.names = FALSE) #note, this line will only work in a markdown file (*.Rmd). It presents the table for a report
 ```
 
@@ -86,16 +79,8 @@ knitr::kable(areaTypes[areaTypes$AreaTypeID == 102,
              row.names = FALSE) #note, this line will only work in a markdown file (*.Rmd). It presents the table for a report
 ```
 
-*ParentAreaTypeID* is 6 by default for the `fingertips_data()` function for `AreaTypeID` of 102 (though it change if different `AreaTypeID`s are entered), so we can stick with that in this example.
+*ParentAreaTypeID* is 6 by default for the `fingertips_data()` function for `AreaTypeID` of 102 (though it changes if different `AreaTypeID`s are entered), so we can stick with that in this example. Use the `area_types()` function to understand more about how areas map to each other.
 
-## Deprivation
-
-We want to plot life expectancy against deprivation information. The package has a `deprivation_deciles()` function that allows us to return this information. This is populated from the Department for Communities and Local Government Indices of Multiple Deprivation (IMD). Note, there is only information for upper and lower tier local authorities (*AreaTypeID* = 102 and 101 respectively). IMD has only been produced for the years 2010 and 2015.
-
-```{r deprivation}
[…]
