### 2018-01-11 09adfb3
stringsAsFactors removed from fingertips_data. Other minor changes to documentation of some functions.
