### 2018-01-11 76dff01
areatypes_by_indicator dataset removed and the call to the API is now used for the indicator_areatypes function
