+++ R/data.R
@@ -1,12 +1,0 @@
-#' IndicatorID by AreaTypeID
-#'
-#' A dataset containing indicator ids and the area types that data exists for
-#'
-#' @format A data frame with 3593 rows and 2 variables:
-#' \describe{
-#'   \item{IndicatorID}{ID for indicator in Fingertips database, see  indicators()}
-#'   \item{AreaTypeID}{ID for area types used in Fingertips database, see area_types()}
-#'   ...
-#' }
-#' @source \url{https://fingertips.phe.org.uk/}
-"areatypes_by_indicators"
