+++ NAMESPACE
@@ -35,5 +35,4 @@ importFrom(shiny,stopApp)
 importFrom(shiny,tableOutput)
 importFrom(shinycssloaders,withSpinner)
 importFrom(stats,complete.cases)
-importFrom(utils,data)
 importFrom(utils,read.csv)
