### 2018-01-11 0a7c121
attempted to remove select_indicators from coverage calculation
