### 2018-01-11 356caa7
updated tests to improve coverage, code of conduct added, NEWs now has release dates, README and DESCRIPTION updated
