+++ tests/testthat/test-area_types.R
@@ -7,6 +7,8 @@ test_that("the area_types function works correctly", {
         expect_is(area_types(), "data.frame")
         expect_equal(nrow(area_types("County")), 0)
         expect_equal(area_types(c("gov","count")), area_types(AreaTypeID = c(102,6)))
+        expect_warning(area_types(AreaTypeName = "gov", AreaTypeID = 102),
+                       "AreaTypeName used when both AreaTypeName and AreaTypeID are entered")
 })
 
 context("Area types by indicator")
@@ -16,5 +18,11 @@ test_that("the indicator_areatypes function is working correctly", {
         expect_equal(ncol(indicator_areatypes(108)), 2)
         expect_equal(ncol(indicator_areatypes(AreaTypeID = 101)), 2)
         expect_error(indicator_areatypes(c(108, 20)), "Length of IndicatorID must be 0 or 1")
-        expect_error(indicator_areatypes(AreaTypeID =c(108, 20)), "Length of AreaType must be 0 or 1")
+        expect_error(indicator_areatypes(AreaTypeID =c(108, 20)), "Length of AreaTypeID must be 0 or 1")
+})
+
+context("category_types works correctly")
+test_that("category_types returns as expected", {
+        expect_is(category_types(), "data.frame")
+        expect_equal(ncol(category_types()), 4)
 })
