+++ NEWS.md
@@ -6,17 +6,17 @@
 
 * package prepared for submission to ropensci
 
-# fingertipsR 0.1.3
+# fingertipsR 0.1.3 (5/10/2017)
 
 * API structure updated to include 99.8 and 95 confidence intervals. Reflected in the outputs of `fingertips_data`. **NOTE** earlier versions of the package will not work anymore because of the underlying change in the API structure
 
-# fingertipsR 0.1.2
+# fingertipsR 0.1.2 (27/9/2017)
 
 * fixed issue with rank and some `fingertips_data` queries
 
 * removed dependency on tidyjson as a result of its removal from CRAN
 
-# fingertipsR 0.1.1
+# fingertipsR 0.1.1 (7/9/2017)
 
 * `select_indicators()` allows user to point and click to select indicators
 
@@ -34,7 +34,7 @@
 
 * A new vignette demonstrating how some of the new functions can be used
 
-# fingertipsR version 0.1.0
+# fingertipsR version 0.1.0 (17/6/2017)
 
 This package allows the user to retrieve tables of:
 
