+++ DESCRIPTION
@@ -3,6 +3,7 @@ Type: Package
 Version: 0.1.3.9000
 Title: Fingertips Data for Public Health
 Description: Fingertips (<http://fingertips.phe.org.uk/>) contains data for many indicators of public health in England. The underlying data is now more easily accessible by making use of the API.
+Maintainer: Sebastian Fox <sebastian.fox@phe.gov.uk>
 Depends: R (>= 2.10)
 Authors@R: c(person("Sebastian", "Fox", , "sebastian.fox@phe.gov.uk", c("aut", "cre")),
     person("Julian", "Flowers", , "julian.flowers@phe.gov.uk", c("aut","ctb")))
