### 2018-01-11 2c6bc5a
all rOpenSci updates from reviewer #1 complete
