+++ man/indicator_areatypes.Rd
@@ -6,11 +6,21 @@
 \usage{
 indicator_areatypes(IndicatorID, AreaTypeID)
 }
+\arguments{
+\item{IndicatorID}{integer; the Indicator ID (can be ignored or of length 1).
+Takes priority over AreaTypeID if both are entered}
+
+\item{AreaTypeID}{integer; the Area Type ID (can be ignored or of length 1)}
+}
 \value{
 A data frame of indicator ids and area type ids
 }
 \description{
-Outputs a data frame of indicator ids and the area type ids that exist for that indicator
+Outputs a data frame of indicator ids and the area type ids that exist for
+that indicator
+}
+\examples{
+indicator_areatypes()
 }
 \seealso{
 \code{\link{indicators}} for indicator lookups,
