### 2018-01-12 601140d
Minor updates based on rOpenSci reviewer #2. About to replace all for loops with lapply
