+++ R/retrieve_data.R
@@ -30,7 +30,12 @@ retrieve_indicator <- function(IndicatorIDs, ProfileIDs, ChildAreaTypeIDs, Paren
                                                                   "&include_sortable_time_periods=yes")
                                         }
                                 }
-                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = types),
+                                fingertips_data <- rbind(dataurl %>%
+                                                                 GET %>%
+                                                                 content("parsed",
+                                                                         type = "text/csv",
+                                                                         encoding = "UTF-8",
+                                                                         col_types = types),
                                                          fingertips_data)
                         }
                 }
@@ -54,7 +59,12 @@ retrieve_domain <- function(DomainIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
                                                   sprintf("all_data/csv/by_group_id?child_area_type_id=%s&parent_area_type_id=%s&group_id=%s",
                                                           ChildAreaTypeID,ParentAreaTypeID,DomainID),
                                                   "&include_sortable_time_periods=yes")
-                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = types),
+                                fingertips_data <- rbind(dataurl %>%
+                                                                 GET %>%
+                                                                 content("parsed",
+                                                                         type = "text/csv",
+                                                                         encoding = "UTF-8",
+                                                                         col_types = types),
                                                          fingertips_data)
                         }
                 }
@@ -78,7 +88,12 @@ retrieve_profile <- function(ProfileIDs, ChildAreaTypeIDs, ParentAreaTypeIDs){
                                                   sprintf("all_data/csv/by_profile_id?child_area_type_id=%s&parent_area_type_id=%s&profile_id=%s",
                                                           ChildAreaTypeID,ParentAreaTypeID,ProfileID),
                                                   "&include_sortable_time_periods=yes")
-                                fingertips_data <- rbind(dataurl %>% GET %>% content("parsed", type = "text/csv", encoding = "UTF-8", col_types = types),
+                                fingertips_data <- rbind(dataurl %>%
+                                                                 GET %>%
+                                                                 content("parsed",
+                                                                         type = "text/csv",
+                                                                         encoding = "UTF-8",
+                                                                         col_types = types),
                                                          fingertips_data)
                         }
                 }
