### 2018-01-12 ropensci/fingertipsR#39 @sebastian-fox
Use `lapply` instead of `for ... loop` in `retrieve_data.R`
Useful feedback from rOpenSci. As the title suggests.
