### 2018-01-12 #168 @sebastian-fox
Hi @noamross , @carlganz  and @nacnudus ,

Firstly, a massive thank you for taking your time to look at this package. I can see you have given it an extremely thorough review and I've learnt **a lot** from your comments, so thank you.

I'll respond to question first, then what I have/haven't addressed:

### Question
Why isn't Fingertips on a .gov.uk website? I'm afraid I don't know the answer (I'm not in any of the teams that would be involved in those discussions). My guess is that it is a legacy issue. PHE was formed relatively recently and Fingertips was popular before then, so there are probably challenges to move it across.

### Addressed

##### General

* All typos have been corrected
* stopped dependency on `purrr` and `data.table` (thanks for the `bind_rows` tip - I've not used that before)
* replace all "for loops" with `lapply` functions except in `retrieve_data.R` file as that's a little more complex (in my eyes at least) as I don't currently have the time. I've added it as an [issue](https://github.com/PublicHealthEngland/fingertipsR/issues/39)
* all "significants" are now preceded by "statistically" for clarity
* code coverage should now be around 100% (it says I'm not testing 1 line when I follow the `covr` instructions @nacnudus  provided, but I'm sure I am testing it!)
* Package tagged on GitHub
* attempted to shorten lines where possible (and where I've spotted them)
* commented out fields have been removed
* tibbles are now returned instead of data.frames
* functions that take time to run now have a comment in the description with that warning

##### README

* Improved explanation of the Fingertips website, its scope, who it is aimed at, what it contains
* Improved package installation instructions
* Code of conduct included
* Example has been improved and the text explaining it should be a little clearer

##### DESCRIPTION

* `Maintainer:` added
* Copyright added

##### NEWS

* CRAN release dates added

##### LE Vignette

* All feedback from both reviewers has been incorporated except the advice not to display one of the tables using `pander::pandoc.table()`. I chose this because I felt this was the clearest way of displaying this dataset with so many fields. I wanted to avoid left-right scrolling, but that was just my style. Any other suggestions are welcomed though

##### Interactively selecting indicators vignette

* `geom_tile` changed to `geom_point` - nice suggestion. I'm not keen to provide a function for that chart at the moment. Perhaps if the package evolves to start displaying standard charts

##### area_types

* added an `AreaTypeID` example
* added reference to the Life Expectancy vignette for further explanation of how areas can map to each other

##### category_types

* example added
* now returns something! (whoops)

##### deprivation_decile

* Details section added
* The years/error messages have been updated - well spotted

##### fingertips_data

* I couldn't see the defaults `6` and `101` there to explain them - apologies if I've overlooked them
* `stringsAsFactors` removed
* Fuller explanation of the IDs in the Life Expectancy vignette. I hope this makes things a little clearer.
* Polarity explained in Details section

##### fingertips_redred

* `@return` renders correctly now
* Text corrected based on different comparator type

##### indicator_areatypes

* function updated to pull data from API (this wasn't available on the API until recently)
* example added

##### indicators

* title corrected

##### select_indicators

* thanks for positive feedback. It was a colleague's idea, I'll pass that onto him :)

##### indicator2

* I can't find this function. I've used "Find in files" but can't see it anywhere. @carlganz - could you point me towards it? It shouldn't exist so I need to delete it.
