### 2018-01-12 7d3fd60
Commit following response to rOpenSci
