+++ R/area_types.R
@@ -1,7 +1,8 @@
 #' Area types
 #'
 #' Outputs a data frame of area type ids, their descriptions, and how they map
-#' to parent area types
+#' to parent area types. To understand more on mappings of areas, see the Where
+#' to start section of the Life Expectancy vignette.
 #'
 #' @return A data frame of area type ids and their descriptions
 #' @param AreaTypeName Character vector, description of the area type; default
