### 2018-01-12 #168 @sebastian-fox
I should have mentioned - all the changes have been made to v0.1.3.9000. I should submit the package to CRAN again soon because of the change in the `indicator_areatypes()` function now pulling data from the API. I'll wait for any further feedback from you before doing so.
