### 2018-01-13 #168 @nacnudus
@sebastian-fox thanks for making all those changes -- that's a lot of work!  There are a couple of small leftovers, but in general I think this is ready to go now.

* Coverage: The one line that isn't covered (159 of `R/fingertips_data.R`) is one that I don't think will ever be executed, because an earlier, equivalent test in the same function will fail first.  It could be tidied up, but I don't think it's a blocker for publication.
* The Travis badge in the README should link to https://travis-ci.org/PublicHealthEngland/fingertipsR to take people who click it to the test results, rather than to the image.
* The new README blurb is very nice.  It has one typo ;) `accesibility` should be `accessibility`.
* The CRAN installation code would render correctly if wrapped as a chunk, rather than as inline code, e.g.
    install.packages("fingertipsR")
* Your use of pandoc in the vignette: fair enough!
* `deprivation_decile()` the test `tests/testthat/test-deprivation.R` needs updating too, and then the build will succeed.  Line 32 should be
```r
        expect_error(deprivation_decile(Year = 2014), "Year must be either 2010, 2011, 2012 or 2015")
````
* The `101`, `102` and `7` defaults have now been explained (perhaps `6` was changed
to `7`?).  Anyway, it makes much more sense to me now, thanks.
