### 2018-01-14 6cc469c
small changes based on further feedback from rOpenSci
