+++ README.md
@@ -1,13 +1,15 @@
 
 <!-- README.md is generated from README.Rmd. Please edit that file -->
-[![Build Status](https://travis-ci.org/PublicHealthEngland/fingertipsR.svg)](https://travis-ci.org/PublicHealthEngland/fingertipsR.svg?branch=master) [![Coverage Status](https://coveralls.io/repos/github/PublicHealthEngland/fingertipsR/badge.svg?branch=master)](https://coveralls.io/github/PublicHealthEngland/fingertipsR?branch=master) [![](https://badges.ropensci.org/168_status.svg)](https://github.com/ropensci/onboarding/issues/168)
+[![Build Status](https://travis-ci.org/PublicHealthEngland/fingertipsR.svg)](https://travis-ci.org/PublicHealthEngland/fingertipsR) [![Coverage Status](https://coveralls.io/repos/github/PublicHealthEngland/fingertipsR/badge.svg?branch=master)](https://coveralls.io/github/PublicHealthEngland/fingertipsR?branch=master) [![](https://badges.ropensci.org/168_status.svg)](https://github.com/ropensci/onboarding/issues/168)
 
 [![CRAN Status Badge](http://www.r-pkg.org/badges/version/fingertipsR)](https://cran.r-project.org/package=fingertipsR) [![CRAN Total Downloads](http://cranlogs.r-pkg.org/badges/grand-total/fingertipsR)](https://cran.r-project.org/package=fingertipsR) [![CRAN Monthly Downloads](http://cranlogs.r-pkg.org/badges/fingertipsR)](https://cran.r-project.org/package=fingertipsR)
 
 fingertipsR
 ===========
 
-This is an R package to interact with Public Health England's [Fingertips](http://fingertips.phe.org.uk/) data tool. Fingertips is a major public repository of population and public health indicators for England. The site presents the information in many ways to improve accesibility for a wide range of audiences ranging from public health professionals and researchers to the general public. This package can be used to load data from the Fingertips API into R for further use.
+This is an R package to interact with Public Health England's [Fingertips](http://fingertips.phe.org.uk/) data tool. Fingertips is a major public repository of population and public health indicators for England. The site presents the information in many ways to improve accessibility for a wide range of audiences ranging from public health professionals and researchers to the general public. The information presented is a mixture of data available from other public sources, and those that are available through user access agreements with other organisations. The source of each indicator presented is available using the `indicator_metadata()` function.
+
+This package can be used to load data from the Fingertips API into R for further use.
 
 Installation
 ------------
@@ -16,7 +18,9 @@ Installation
 
 Get the latest released, stable version from CRAN:
 
-`{ r CRAN install, eval=FALSE} install.packages("fingertipsR")`
+``` r
+install.packages("fingertipsR")
+```
 
 ### With devtools
 
@@ -60,20 +64,15 @@ library(fingertipsR)
 profs <- profiles()
 profs <- profs[grepl("Public Health Outcomes Framework", profs$ProfileName),]
 head(profs)
-#>    ProfileID                      ProfileName   DomainID
-#> 22        19 Public Health Outcomes Framework    1000049
-#> 23        19 Public Health Outcomes Framework    1000041
-#> 24        19 Public Health Outcomes Framework    1000042
-#> 25        19 Public Health Outcomes Framework    1000043
-#> 26        19 Public Health Outcomes Framework    1000044
-#> 27        19 Public Health Outcomes Framework 1938132983
-#>                            DomainName
-#> 22             Overarching indicators
[…]
