### 2018-01-14 #168 @sebastian-fox
Hi @nacnudus ,

Thanks for the further comments - all really helpful again.
I've made all the changes now:

* removed line 159 from `fingertips_data()` as it was being tested on line 91
* updated link to travis
* corrected accessibility typo
* fixed the CRAN install README chunk to render properly
* corrected the `deprivation_decile()` test

And this time I've checked that the package builds.

Thanks again
