### 2018-01-18 10dabe3
Removing geom_tile() from text
