### 2018-01-18 #168 @sebastian-fox
Thanks all - I can see a lot of work goes into this from your end so this is really appreciated.

I will transfer the package over as soon as I can. I'm don't have admin access currently to my repo so I can't see the settings option. I need to negotiate admin rights, so hopefully that won't take too long.

Thanks again
