+++ R/area_types.R
@@ -41,9 +41,7 @@ area_types  <- function(AreaTypeName = NULL, AreaTypeID = NULL, path){
         if (!(is.null(AreaTypeName)) & !(is.null(AreaTypeID))) {
                 warning("AreaTypeName used when both AreaTypeName and AreaTypeID are entered")
         }
-        if (missing(path)) {
-                path <- "https://fingertips.phe.org.uk/api/"
-        }
+        if (missing(path)) path <- "https://fingertips.phe.org.uk/api/"
         set_config(config(ssl_verifypeer = 0L))
         parentAreas <- paste0(path,"area_types/parent_area_types") %>%
                 GET %>%
@@ -102,9 +100,7 @@ area_types  <- function(AreaTypeName = NULL, AreaTypeID = NULL, path){
 #'   \code{\link{indicators_unique}} for unique indicatorids and their names
 
 category_types <- function(path) {
-        if (missing(path)) {
-                path <- "https://fingertips.phe.org.uk/api/"
-        }
+        if (missing(path)) path <- "https://fingertips.phe.org.uk/api/"
         set_config(config(ssl_verifypeer = 0L))
         category_types <- paste0(path,"category_types") %>%
                 GET %>%
@@ -140,9 +136,7 @@ category_types <- function(path) {
 #'   \code{\link{category_types}} for category type lookups and
 #'   \code{\link{indicators_unique}} for unique indicatorids and their names
 indicator_areatypes <- function(IndicatorID, AreaTypeID, path) {
-        if (missing(path)) {
-                path <- "https://fingertips.phe.org.uk/api/"
-        }
+        if (missing(path)) path <- "https://fingertips.phe.org.uk/api/"
         path <- paste0(path, "available_data")
         if (!missing(IndicatorID)) {
                 if (length(IndicatorID) > 1) {
