+++ R/profiles.R
@@ -27,9 +27,7 @@
 #' @export
 
 profiles <- function(ProfileID = NULL, ProfileName = NULL, path) {
-        if (missing(path)) {
-                path <- "https://fingertips.phe.org.uk/api/"
-        }
+        if (missing(path)) path <- "https://fingertips.phe.org.uk/api/"
         set_config(config(ssl_verifypeer = 0L))
         profiles <- paste0(path,"profiles") %>%
                 GET %>%
