### 2018-01-19 ddcf6c3
changed case for badge links in README
