### 2018-01-19 f9545b7
added README footer
