+++ README.md
@@ -127,3 +127,5 @@ More
 
 -   Please note that this project is released with a [Contributor Code of Conduct](CONDUCT.md). By participating in this project you agree to abide by its terms
 -   License: [GPL-3](https://opensource.org/licenses/GPL-3.0)
+
+[![ropensci\_footer](https://ropensci.org/public_images/ropensci_footer.png)](https://ropensci.org)
