### 2018-01-19 #168 @sebastian-fox
Hi @noamross - I've completed your checklist. Thanks for your help. I'll start thinking about blogs :)
