### 2018-04-20 #168 @sebastian-fox
Hello, just wondering how to get the rOpenSci badge status updated from "Under review" to "peer reviewed" for `fingertipsR` on the README file?
